requirements.txt
psycopg2-binary==2.9.11
python-dotenv==1.0.0
supabase==2.22.2